Refactor ateom networking to shared internal/ateomnet package
- Extracts duplicated networking logic from ateom-gvisor and ateom-microvm into internal/ateomnet.
- Updates gVisor CleanupActorNetwork behavior so that an error from RemoveActorNftablesRules logs a warning and continues cleanup rather than failing fast.
